[svn-r6713] Purpose:     Buglet fix
Description:
    The 'rdcc_nelmts' parameter to H5P[set|get]_cache is an int, not a size_t.

Solution:
    Changed it. :-)

Platforms tested:
    FreeBSD 4.8 (sleipnir)
    (Too minor to need h5committest)

+++ test/unlink.c
@@ -395,7 +395,7 @@ main(void)
 
     /* Metadata cache parameters */
     int mdc_nelmts;
-    size_t rdcc_nelmts;
+    int rdcc_nelmts;
     size_t rdcc_nbytes;
     double rdcc_w0;
